requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
chromadb
fastapi
uvicorn
beautifulsoup4
pydantic
python-dotenv
tiktoken
sentence-transformers
requests